requirements.txt
python-telegram-bot>=20.6
Flask>=2.0
aiohttp>=3.8

